+++ src/Console/LaraLensCommand.php
@@ -8,13 +8,18 @@ use HiFolks\LaraLens\LaraLensServiceProvider;
 
 class LaraLensCommand extends Command
 {
-    protected $signature = 'laralens:diagnostic';
+    protected $signature = 'laralens:diagnostic
+                            {op=overview : What you want to see, overview or allconfigs}';
 
     protected $description = 'Show some application configruation.';
 
-    public function handle()
+    private function allConfigs()
     {
-        //$this->info("Start");
+        $this->info(json_encode(config()->all(), JSON_PRETTY_PRINT));
+    }
+
+    private function overview()
+    {
         $ll = new LaraLens();
         $output = $ll->getConfigs();
         $this->table(["Configs", "Values"], $output->toArray());
@@ -28,6 +33,24 @@ class LaraLensCommand extends Command
         $this->table(["Database", "Values"], $output->toArray());
 
         $this->call('migrate:status');
+
+    }
+    public function handle()
+    {
+        $op = $this->argument("op");
+        switch ($op) {
+            case 'overview':
+                $this->overview();
+                break;
+            case 'allconfigs':
+                $this->allConfigs();
+                break;
+
+            default:
+                $this->info("What you mean? try with 'php artisan laralens:diagnostic --help'");
+                break;
+        }
+        //$this->info("Start");
 
     }
 }